requirements.txt
tk
pyscreenshot
pyautogui
pillow
pytesseract
pygetwindow
opencv-python
keyboard
